Memosテーブルに関する処理を別のクラスに切り出し
